Add ability to click and see new research
